python-backend/requirements.txt
fastapi
uvicorn
openai==1.100.2
supabase
python-dotenv
httpx==0.27.2